# schablone.gen.go
// Package Swagger provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.9.0 DO NOT EDIT.
package Swagger

import (
	"fmt"
	"net/http"

	"github.com/deepmap/oapi-codegen/pkg/runtime"
	"github.com/labstack/echo/v4"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET category/create)
	GetCategoryCreate(ctx echo.Context, params GetCategoryCreateParams) error

	// (GET category/get/{categoryId})
	GetCategoryGetCategoryId(ctx echo.Context, categoryId int) error

	// (GET category/list)
	GetCategoryList(ctx echo.Context, params GetCategoryListParams) error

	// (GET template/create)
	GetTemplateCreate(ctx echo.Context, params GetTemplateCreateParams) error

	// (GET template/edit)
	GetTemplateEdit(ctx echo.Context, params GetTemplateEditParams) error

	// (GET template/list)
	GetTemplateList(ctx echo.Context, params GetTemplateListParams) error

	// (GET user/create)
	GetUserCreate(ctx echo.Context, params GetUserCreateParams) error

	// (GET user/get/{userId})
	GetUserGetUserId(ctx echo.Context, userId int) error

	// (GET user/list)
	GetUserList(ctx echo.Context, params GetUserListParams) error

	// (GET user/set/{userId})
	GetUserSetUserId(ctx echo.Context, userId int) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetCategoryCreate converts echo context to params.
func (w *ServerInterfaceWrapper) GetCategoryCreate(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{""})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetCategoryCreateParams
	// ------------- Required query parameter "name" -------------

	err = runtime.BindQueryParameter("form", true, true, "name", ctx.QueryParams(), &params.Name)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter name: %s", err))
	}

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.GetCategoryCreate(ctx, params)
	return err
}

// GetCategoryGetCategoryId converts echo context to params.
func (w *ServerInterfaceWrapper) GetCategoryGetCategoryId(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "categoryId" -------------
	var categoryId int

	err = runtime.BindStyledParameterWithLocation("simple", false, "categoryId", runtime.ParamLocationPath, ctx.Param("categoryId"), &categoryId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter categoryId: %s", err))
	}

	ctx.Set(ApiKeyAuthScopes, []string{""})

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.GetCategoryGetCategoryId(ctx, categoryId)
	return err
}

// GetCategoryList converts echo context to params.
func (w *ServerInterfaceWrapper) GetCategoryList(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{""})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetCategoryListParams
	// ------------- Required query parameter "categoryId" -------------

	err = runtime.BindQueryParameter("form", true, true, "categoryId", ctx.QueryParams(), &params.CategoryId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter categoryId: %s", err))
	}

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.GetCategoryList(ctx, params)
	return err
}

// GetTemplateCreate converts echo context to params.
func (w *ServerInterfaceWrapper) GetTemplateCreate(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{""})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetTemplateCreateParams
	// ------------- Required query parameter "name" -------------

	err = runtime.BindQueryParameter("form", true, true, "name", ctx.QueryParams(), &params.Name)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter name: %s", err))
	}

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.GetTemplateCreate(ctx, params)
	return err
}

// GetTemplateEdit converts echo context to params.
func (w *ServerInterfaceWrapper) GetTemplateEdit(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{""})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetTemplateEditParams
	// ------------- Required query parameter "name" -------------

	err = runtime.BindQueryParameter("form", true, true, "name", ctx.QueryParams(), &params.Name)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter name: %s", err))
	}

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.GetTemplateEdit(ctx, params)
	return err
}

// GetTemplateList converts echo context to params.
func (w *ServerInterfaceWrapper) GetTemplateList(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{""})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetTemplateListParams
	// ------------- Required query parameter "groupId" -------------

	err = runtime.BindQueryParameter("form", true, true, "groupId", ctx.QueryParams(), &params.GroupId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter groupId: %s", err))
	}

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.GetTemplateList(ctx, params)
	return err
}

// GetUserCreate converts echo context to params.
func (w *ServerInterfaceWrapper) GetUserCreate(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{""})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetUserCreateParams
	// ------------- Required query parameter "userId" -------------

	err = runtime.BindQueryParameter("form", true, true, "userId", ctx.QueryParams(), &params.UserId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter userId: %s", err))
	}

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.GetUserCreate(ctx, params)
	return err
}

// GetUserGetUserId converts echo context to params.
func (w *ServerInterfaceWrapper) GetUserGetUserId(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "userId" -------------
	var userId int

	err = runtime.BindStyledParameterWithLocation("simple", false, "userId", runtime.ParamLocationPath, ctx.Param("userId"), &userId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter userId: %s", err))
	}

	ctx.Set(ApiKeyAuthScopes, []string{""})

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.GetUserGetUserId(ctx, userId)
	return err
}

// GetUserList converts echo context to params.
func (w *ServerInterfaceWrapper) GetUserList(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{""})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetUserListParams
	// ------------- Required query parameter "userId" -------------

	err = runtime.BindQueryParameter("form", true, true, "userId", ctx.QueryParams(), &params.UserId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter userId: %s", err))
	}

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.GetUserList(ctx, params)
	return err
}

// GetUserSetUserId converts echo context to params.
func (w *ServerInterfaceWrapper) GetUserSetUserId(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "userId" -------------
	var userId int

	err = runtime.BindStyledParameterWithLocation("simple", false, "userId", runtime.ParamLocationPath, ctx.Param("userId"), &userId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter userId: %s", err))
	}

	ctx.Set(ApiKeyAuthScopes, []string{""})

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.GetUserSetUserId(ctx, userId)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"category/create", wrapper.GetCategoryCreate)
	router.GET(baseURL+"category/get/:categoryId", wrapper.GetCategoryGetCategoryId)
	router.GET(baseURL+"category/list", wrapper.GetCategoryList)
	router.GET(baseURL+"template/create", wrapper.GetTemplateCreate)
	router.GET(baseURL+"template/edit", wrapper.GetTemplateEdit)
	router.GET(baseURL+"template/list", wrapper.GetTemplateList)
	router.GET(baseURL+"user/create", wrapper.GetUserCreate)
	router.GET(baseURL+"user/get/:userId", wrapper.GetUserGetUserId)
	router.GET(baseURL+"user/list", wrapper.GetUserList)
	router.GET(baseURL+"user/set/:userId", wrapper.GetUserSetUserId)

}