# types.gen.go
// Package Swagger provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.9.0 DO NOT EDIT.
package Swagger

const (
	ApiKeyAuthScopes = "ApiKeyAuth.Scopes"
)

// Template defines model for Template.
type Template struct {
	AttachementIds *[]int  `json:"attachementIds,omitempty"`
	Content        *string `json:"content,omitempty"`
	Id             *int    `json:"id,omitempty"`
	Title          *string `json:"title,omitempty"`
}

// GetCategoryCreateParams defines parameters for GetCategoryCreate.
type GetCategoryCreateParams struct {
	// Category title
	Name string `json:"name"`
}

// GetCategoryListParams defines parameters for GetCategoryList.
type GetCategoryListParams struct {
	// The category ID
	CategoryId int `json:"categoryId"`
}

// GetTemplateCreateParams defines parameters for GetTemplateCreate.
type GetTemplateCreateParams struct {
	// Category title
	Name string `json:"name"`
}

// GetTemplateEditParams defines parameters for GetTemplateEdit.
type GetTemplateEditParams struct {
	// Category title
	Name string `json:"name"`
}

// GetTemplateListParams defines parameters for GetTemplateList.
type GetTemplateListParams struct {
	// Group ID
	GroupId int `json:"groupId"`
}

// GetUserCreateParams defines parameters for GetUserCreate.
type GetUserCreateParams struct {
	// The user ID
	UserId int `json:"userId"`
}

// GetUserListParams defines parameters for GetUserList.
type GetUserListParams struct {
	// The user ID
	UserId int `json:"userId"`
}